Fix isimodem segfault on modem removal
This fixes segfaults resulting either from destroying a NULL ISI
client, or when multiple ISI modems become available, and get removed
in the wrong order. To support multiple ISI modems concurrently, a
modem object needs to be exposed via the gisi API, and tracked by the
ISI driver. This is a TODO item to add.