Fix stupid false positive warnings about uninitialized variables
The GCC check for uninitialized variables is so useful that dealing
with the false positives is just a minor inconvenience.

The following GCC trick helps to silence them:

	#define uninitialized_var(x) x = x

This way no code obfuscation is needed to compile smoothly and the
code stays readable. And the author has to actually think about it.