bootstrap: autoconf must run before automake
The other way, dependencies computation (depcomp) does not work
correctly. In fact, aclocal is the part of GNU/automake to be run
before autoconf whereas automake is the part to be run after it.
In other words, aclocal would not exist if automake could run
properly before autoconf. This ordering is consistent with most
autotools-based projects, and most authoritatively, with autoreconf.

This also adds a missing anti-slash for consistency.