hfp: set timeout to rfcomm establishment procedure
If it expires we call org.bluez.HandsfreeGateway.Disconnect()
It also modifies send_method_call and break it down into two functions.
Now we have send_method_call_with_reply, for DBus messages that needs a
reply. It sets a reply and a timeout for the message reply.
The other funcion is called send_method_call without timeout or reply
parameters.