Change the DataConnectionManager API
The various attributes obtained from CGREG are actually completely
useless and are not used internally for anything except computing the
Attached status.  It makes sense to get rid of these properties since
they will be the same as reported by CREG in Dual-Transfer-Mode devices.
The only difference might be how hardware reports the technology used,
but frequently a vendor specific command is required anyway