stkutil: Use struct to represent frame id
Use a field has_id to indicate if we have frame id or not. Originally
we use 0xFF to indicate we don't have frame id. But in that way, we
have to initialize each test case of proactive command when it doesn't
have this comprehension tlv to be 0xFF. To use has_id will save this
effort, and keep consistency with other implementations.