atmodem: Add really nasty hack for Nokia AT+COPS bug
The Nokia modem returns an invalid response when trying to read the
numeric value for MCC/MNC of the current network.

Control:> AT+COPS?\r
Control:< \r\n+COPS: 0,2,"26207,0\r\n\r\nOK\r\n

The closing " is just missing and this confuses the AT parser. In case
of a Nokia modem skip reading the numerical values for the network.