ifxmodem: Enable SIM Toolkit with AT+CFUN=6
The current Infineon firmware overloads the AT+CFUN command with
enabling SIM Toolkit and fetching of proactive commands. It seems
to be safe to send this command in offline mode (AT+CFUN=4) and
also in online mode (AT+CFUN=1). These are the two cases where
SIM Toolkit atom would be initialized.