ifx: Wait for DLC creation to settle (internal multiplexer)
When driving the Infineon modem with the builtin multiplexer there is
a small race condition with setting up the channels and sending the
first AT commands. The window here is pretty small, but it seems to be
a modem firmware issue. In case the AT command is send right away it can
happen that the modem does not process any further AT commands. In that
case the setup is stuck and enabling the modem fails.

Just adding a 10 milliseconds delay after DLC creation and before
sending the first AT commands is enough to make this work smoothly.