ifxmodem: Signal STK session end only after normal ending of command
The Infineon modem returns more detailed information of the SIM Toolkit
proactive sessions to the modem driver. A proactive session starts with
the status response '91 xx' (proactive command pending) and ends with a
status response response of '90 00' (normal ending of command) after
Terminal Response.

So only when receiving '90 00' result, the indication that the session
has ended should be given to the core.