include: define initializer for ofono_call
Some fields in struct ofono_call such as cnap_validity and clip_validity
are not correctly initialized as 0 since this means they are present. If
driver does not implement the necessary handlers, it might send to core
an ofono_call signalling that cnap and clip are available when they are
actually not.