bluetooth: Refactor the authorization process
It is unnecessary to keep the IO watch around after the authorization
process has finished.  This also means client_list variable is not
really needed.

Since we destroy the watch once the auth_cb is called, we can also
remove the auth_pending variable.

Finally, the null check of path is not necessary