modem: Fix entering offline mode prematurely
A contextual change to __ofono_modem_find_atom has caused us to enter
into offline mode prematurely.  find_atom was changed to return only the
atoms that were registered, however the modem really has to know whether
the sim atom is created (not registered) in order to know whether to
wait for the SIM to be initialized or not.