network: Split EFspn and EFspdi Refresh handling
If both EFspn and EFspdi are changed, then we trigger reading of EFspn
twice which leads to a memory leak.  Instead, always read EFspdi if the
relevant service is available.

If EFspdi is changed, use a simple heuristic to update the 'Name'
property if appropriate.  This heuristic is not always correct, but in
the worst case we will emit the same name.