ifx: Bump the setup timeout
The newer IFX firmware was changed, it no longer accepts AT commands
until the UA frame has been sent back to the AP.  The commands are
simply dropped on the floor.  This requires a proper fix in GAtMux, but
until then simply bump up the timeout to get around this issue.