rilmodem: Fix GPRS attach/detach logic ( LP: 1234434 ).
When the primary context was shutdown, the active_ctx_cid was
being reset before rilmodem notified the core that the context
had been deactified.  This meant an invalid (-1) cid was passed
to the core, and thus the context wasn't being cleaned up properly,
which in turn could cause the attach code to quit responding to
subsequent attaches ( either from DBus requests, or DATA
registration changes.