mtkmodem: New mtk modem driver
This driver provides support for MediaTek modems. Most of the
functionality is already provided by the rilmodem driver: this driver
provides support for specific MTK parcels and voicecall and gprs atoms.
These atoms are separated from those in rilmodem as they use specific
MediaTek requests and events. This will allow us to move the driver to
a dynamically loaded plugin if desired.