mtk: Resume registration when needed
Arale produces REGISTRATION_SUSPENDED events that need to be answered
with a RESUME_REGISTRATION command to resume registration. These events
are generated jointly with PLMN_CHANGED events and are probably used
to stop registration and let rild client check if the modem has a valid
FW for the operator's technology. This kind of behaviour will be
supported in future commits.