mtkmodem: Fix failure to answer fast a call
When an incoming call was being answered immediately after it appearead,
RIL_REQUEST_ANSWER was sent and the reply indicated success, but the
call was not actually transitioning to the active state. This is a bug
in rild because the modem was responding with error "+CME: 3" to the
"ATA" command, but nonetheless there is a also bug in ofono because we
are letting the user answer the call before the modem is ready. To avoid
this, we do not poll for the calls after sending an
MTK_CALL_INDIC_MODE_AVAIL, but wait for the
MTK_RIL_UNSOL_CALL_PROGRESS_INFO that indicates that the modem is ready
to do so.