gprs: Resolve MMS proxy/MMSC host name
Resolve, in case the hostname is provided instead of the IP address,
MMS proxy / MMSC name using the DNS servers provided by the context.
We need the IP so we can set-up a route for them. Also, use the MMSC
for the route in case we do not have a MMS proxy. Finally, export proxy
and proxy port as part of the Settings property array, for both combined
and mms context types. Client must use these values to connect to make
sure the used IP is the same one as the one for the route, as the MMSC
might be a balanced server with multiple IPs.